pkgfile module for summit
pkgfile module provides all functionalities to add/delete/rollback any entries in roxie package file which can then be used to deploy it to roxie. This is alternative for packagemap which is a command line tool. The module provides a greater flexibility for ECL users such that they can incorporate this module in their data build code and update the package file automatically. All the needed information will be stored in a flat file and whenever a package file is needed (which is xml) the information is pulled from flat and converted to xml in the needed format.