HPCC-15598 Suppress buildindex,width(1) warning
The performance suite is specifically testing the building of a
large 1-way index. This normally generates a warning to warn the
user that they may not have intended to build such a large single
part index on Thor.
In the performace suite this warning only serves to cause the
result to mismatch the result from other engines and the key
expected.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>